+++ CHANGELOG.md
@@ -7,13 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### Added
-* `disabled` key for readme to avoid including demos in project readme
+- `excludeFromModules` key for readme to avoid including demos in project readme
 
 ### Changed
+- Features are moved to `docs/features.json`
 
 ### Deprecated
+- Features inside the build file are deprecated
 
 ### Removed
+- js plugin
 
 ### Fixed